+++ domain/src/main/java/com/muhmmad/domain/model/ActivityResponse.kt
@@ -3,7 +3,7 @@ package com.muhmmad.domain.model
 data class ActivityResponse(
     val book: Book,
     val createdAt: String,
-    val readingProgress: Int,
+    val readingProgress: Double,
     val status: String,
     val updatedAt: String
 )
